requirements.txt
fastapi
uvicorn
pydantic
estnltk
python-multipart
httpx
pydantic[email]
sqlalchemy
asyncpg
psycopg2-binary
pyjwt
bcrypt
sqlmodel
aiofiles
aiobotocore